Be/#112

+++ BE/classes/views.py
@@ -1,4 +1,5 @@
 from django.http import JsonResponse
+from lectures.permissions import IsLectureMember
 from classes.tasks import run_speech, save_temp_audio
 from classes.serializers import *
 from classes.models import Bookmark, Note, Speech
@@ -12,14 +13,14 @@ from rest_framework.views import APIView
 
 
 class SpeechView(APIView):
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = [permissions.IsAuthenticated, IsLectureMember]
 
     def post(self, request, pageId):
         try:
             page = Page.objects.get(id=pageId)
         except Page.DoesNotExist:
             return Response({"error": "해당 페이지를 찾을 수 없습니다."}, status=404)
-        
+        self.check_object_permissions(request, page)
         serializer = SpeechCreateSerializer(data=request.data)
         serializer.is_valid(raise_exception=True)
 
@@ -75,14 +76,14 @@ class TTSTestView(APIView):
             )
         
 class BookmarkView(APIView):
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = [permissions.IsAuthenticated, IsLectureMember]
 
     def post(self, request, pageId):
         try:
             page = Page.objects.get(id=pageId)
         except Page.DoesNotExist:
-            return JsonResponse({"error": "해당 페이지를 찾을 수 없습니다."}, status=404)\
-            
+            return JsonResponse({"error": "해당 페이지를 찾을 수 없습니다."}, status=404)
+        self.check_object_permissions(request, page)    
         timestamp = request.data['timestamp']
 
         if not timestamp:
@@ -104,14 +105,14 @@ class BookmarkView(APIView):
         }, status=200)
     
 class BookmarkDetailView(APIView):
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = [permissions.IsAuthenticated, IsLectureMember]
 
     def get(self, request, bookmarkId):
         try:
             bookmark = Bookmark.objects.get(id=bookmarkId)
         except Bookmark.DoesNotExist:
             return JsonResponse({"error": "해당 북마크를 찾을 수 없습니다."}, status=404)
-        
+        self.check_object_permissions(request, bookmark)
         # 해당 북마크와 매칭되는 Speech 찾기
         speeches = Speech.objects.filter(page=bookmark.page)
         matched_speech = next(
@@ -135,19 +136,19 @@ class BookmarkDetailView(APIView):
             bookmark = Bookmark.objects.get(id=bookmarkId)
         except Bookmark.DoesNotExist:
             return JsonResponse({"error": "해당 북마크를 찾을 수 없습니다."}, status=404)
-        
+        self.check_object_permissions(request, bookmark)
         bookmark.delete()
         return JsonResponse({"message": "북마크가 삭제되었습니다."}, status=200)
 
 class NoteView(APIView):
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = [permissions.IsAuthenticated, IsLectureMember]
 
     def post(self, request, pageId):
         try:
             page = Page.objects.get(id=pageId)
         except Page.DoesNotExist:
             return Response({"error": "해당 페이지를 찾을 수 없습니다."}, status=404)
-
+        self.check_object_permissions(request, page)
         user = request.user
         content = request.data.get("content")
 
@@ -163,7 +164,7 @@ class NoteView(APIView):
         return Response(NoteSerializer(note).data, status=201)
     
 class NoteDetailView(APIView):
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = [permissions.IsAuthenticated, IsLectureMember]
 
     def patch(self, request, noteId):
         note = Note.objects.filter(id=noteId).first()
@@ -172,7 +173,7 @@ class NoteDetailView(APIView):
 
         if note.user != request.user:
             return Response({"error": "본인이 작성한 노트만 수정할 수 있습니다."}, status=403)
-        
+        self.check_object_permissions(request, note)
         content = request.data.get("content", "").strip()
         note.content = content
